build(vscode): update spell checker settings
This commit updates the settings.json file with new settings
for the spell checker integrated in VS Code.

The spell checker will now check for spelling errors in English (en),
British English (en-gb), US English (en-us), and Lorem Ipsum (lorem) language.

In addition, the word "Paradiso" has been added to the list of
words to ignore in order to avoid false errors as the word is in Italian.