feat: Add package.json for npm and SASS integration
This commit adds the package.json file to the project using the npm init command.
This file contains information about the project and its dependencies,
and is necessary for using npm packages.

The package.json file has been added to prepare the project for SASS integration in the future.
SASS is a popular CSS preprocessor that extends CSS with features like variables,
nesting, and functions, and can be installed as an npm package.