feat: Add npm script for SASS compilation
This commit adds an npm script called sass to the package.json file.
This script compiles the main.scss file in the scss folder and outputs
the resulting compressed CSS file into the css folder.

The sass script also watches the scss folder for changes and automatically
recompiles the CSS file when changes are detected.