feat: Add responsive styles for ".container" using flexbox properties
This commit adds responsive styles to the ".container" class,
including background color, font color, margin, and padding.
The styles have been implemented using Flexbox properties to ensure responsiveness,
with the ".container" class becoming 1200 pixels wide when there is
enough space and shrinking when the available space is limited,
such as on mobile devices.