chore: Make minor adjustments and fixes to header section
This commit makes several minor adjustments and fixes to the header section.

Firstly, the logo has been made clickable and now brings the user back
to the index.html page.

Secondly, the width of the logo is now editable through a new variable
added to the _variables.scss partial file called: --logo-width.

Thirdly, the font of the navigation menu text is now editable by changing
the value of the --nav-text-font variable in the _variables.scss file.

Fourthly, the logo now has a class named "header__logo-img" which is
more specific than selecting the img tag itself.

Fifthly, a background color has been added to the header section for
when it scrolls, and this color is now editable through a variable
in the _variables.scss partial called: --header-background-color.

Finally, a new variable called --spacing-xxxs has been added to provide
more tuning options, and it has been set to 0.26rem, which is mostly
used for letter-spacing properties.

With these changes, the header section is now more functional and
customizable, and ready for further development.