feat(hero-section): Create "_hero.scss" partial file and style
This commit completes the hero section of the project and separates it
into a partial file called "_hero.scss". The hero section now includes
a slogan and an h1 title, and it is ready for further development.

The button for the hero section must still be completed, now is
just a placeholder.