feat(css-pattern): Add "abstract" folder and "_buttons" partial for 7-in-1 CSS pattern
This commit introduces the first step in implementing the 7-in-1 CSS pattern.
A new folder called "abstract" has been added to the project, and a new
partial file named "_button" has been created inside this folder to
contain the styles for buttons.

This change will make it easier to manage and maintain the CSS code by
separating it into smaller, more manageable parts.

In addition, this commit marks a gradual shift from using @import to @use
in the SCSS files, which is the recommended approach for importing
partials in newer versions of Sass.

Finally, the variables in the "_variables" partial are being renamed to
more concise and appropriate names.