feat(css-pattern): Create "base" folder and move reset partial to it
This commit continues the implementation of the 7-in-1 CSS pattern by
creating a new "base" folder in the project and moving the reset
partial file into it.

This change helps to organize the project's CSS code by separating the
foundational styles from other styles.

Additionally, this commit marks another step in the gradual transition
from using @import to @use in the SCSS files.