refactor(variables): Clean up and rename variables in _variables partial
This commit cleans up the _variables partial file in the project and
renames several poorly named variables to more appropriate and concise names.

The file has also been cleaned up by removing old comments that are
no longer relevant. By doing this, the code in the file is more organized
and easier to understand, making it simpler to make changes to the project
in the future.