refactor(css): Fixed container class to be more reusable with a flex modifier
This commit improves the container class in the project's CSS by adding
a flex modifier that allows it to be more reusable.

Previously, the container class was limited to a specific flex direction,
either row or column. However, with this change, a modifier has been added
that allows the container class to be easily modified between the two
flex directions.